Reorganize and comment find_package layout
Comment out find_package for gccxml because this is used behind
the scenes for dictionary generation and is thus not a direct
dependency.

Add ART_ENABLE_TESTING option, OFF by default, to activate unit
testing. If ON, enable CTest testing, find CPPUnit and recurse into
test subdirectory. Relocation of find_package for CPPUnit done
because this is only needed for testing.