Build art_Framework_Principal library
Provide minimal build of art_Framework_Principal library. Identified
as a low level library in the Framework subsystem, with minimal
dependency inside that subsystem. Add library and define include/link
interfaces,installing headers and library to standard locations.

Use full link interface as specified in old CMake scripts. Noted that
CLHEP appears not to be used.

Document README, including note that compilation of the library
requires support for C++11 ref-qualified member functions (via
use of `PtrVector.h`). This is not checked for, and is only
supported on GCC 4.8 and above (Clang support not fully determined).