Implement support for checkClassVersion in art_dictionary
Add '-L' command line option to checkClassVersion to allow
extra paths to be supplied for dynamic lookup. This avoids the
requirement for a (DY)LD_LIBRARY_PATH that includes the build
directory (both for art and clients).

Provide function in CheckClassVersion CMake module to add
paths to checkClassVersion via a global property. A property
is used because the call chain of art_dictionary and CheckClassVersion
makes it difficult to add these paths via arguments to those functions.

Replace full path to checkClassVersion call in add_custom_command
with an imported target that wraps the checkClassVersion script.
Set up imported target at top level for art itself, and provide
an imported target in artConfig.cmake for clients. This allows
transparent use of the program via an imported target in both art
itself and clients.

Install checkClassVersion program to CMAKE_INSTALL_BINDIR for now.

KNOWN ISSUES: ROOT must be configured with appropriate paths, namely
(DY)LD_LIBRARY_PATH and PYTHONPATH if needed by the install pattern.
No checks are performed yet in art of the correct functioning of
checkClassVersion prior to build starting.