Provide basic CMake tools for art devs/users
Provide two CMake modules to help developers and users apply standard
art patterns for building libraries and plugins. Divided into:

- artInternalTools.cmake for use in developing art only
  - Function to set standard art-wide library properties
  - Variables to provide standard install locations
- artTools.cmake for use by devs and users
  - Functions to create service, module and source plugins. Implement
    as lightweight wrappers around add_library and target_link_libraries
    to check/enforce naming convention for plugins and link to the minimum
    required libraries.

Whilst these somewhat follow the functionality in cetbuildtools, some
significant ease-of-use improvements are made:

- No assumptions are made about locations/names of sources.
- Choice of further linking/installation is left up to dev/user

The intent is to remove many of the hardcoded assumptions in the
current cetbuildtools implementations and locate functionality where
it's needed (i.e. stuff relating to art in art).