Add unit tests for art/Framework/EventProcessor
Recurse into tests directory for art/Framework/EventProcessor. Update
use of Statemachine_t.sh so that it can be used directly without uneeded
copy to binary dir (which must then be in the PATH).

Note that tests *will fail* due to f****d up usage of direct source
of a cetbuildtools shell script. That will be resolved later with a
proper CMake package for that.