Use cet_test_functions script in StateMachine tests
StatteMachine tests are run through a shell script. This sources
the "cet_test_functions.sh" script which is expected to be in the
PATH. The script is supplied as part of cetbuildtools, but implements
functionality referring to art.

Import cet_test_functions.sh into art's Modules dir as is refers to
art, so is part of art. Add this directory to the cet_test environment
as part of the PATH. This ensures that other tests within art that use
it (such as Integration) can use this.