Update CetTest with filtering output
Import CetTest and support cmake and perl script for running/comparing
test output with REF argument.

Whilst this is a "hack", note that filter-output is specialized to
work with (and thus dependent on) messagefacility output. Also not
clear how reference files are created in the first place. There is
therefore a nasty amount of coupling between supposedly low level
cetbuildtools and the higher level systems of messagefacility,
art and mrb...