Build/Install Art Workbook
Build and install the develop branch of the Art workbook as an
external project. Use pdflatex to compile the two pdfs and install
these to the documents folder.

Use ExternalProject until documentation is properly unified.

As the FNAL redmine service is very slow off-site, the git clone
step f the external project may take some time.